+++ docs/nice_to_have.md
@@ -1,0 +1,3 @@
+# Nice to have
+
+- Integrate filtering/sorting as needed.
