+++ src/helpers/types.ts
@@ -1,0 +1,2 @@
+export type TODO = any;
+export type TShirt = "XS" | "S" | "M" | "L" | "XL";
